requirements.txt
scikit-learn
nltk
pytesseract
openai
python-dotenv
sentence_transformers